+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## v2.14.0-beta.1
+
+### Added
+
+ * `Contract[Create|Update]Transaction.[get|set]autoRenewAccountId`
+ * `ContractDeeteTransaction.[get|set]PermanentRemoval`
+
 ## v2.12.0
 
 ### Added